moving all exceptions to root
