+++ tests/soupsavvy/tags/attributes/class_selector_test.py
@@ -4,8 +4,8 @@ import re
 
 import pytest
 
+from soupsavvy.exceptions import TagNotFoundException
 from soupsavvy.tags.attributes import ClassSelector
-from soupsavvy.tags.exceptions import TagNotFoundException
 from tests.soupsavvy.tags.conftest import find_body_element, strip, to_bs
