+++ tests/soupsavvy/tags/combinators/descendant_combinator_test.py
@@ -2,6 +2,7 @@
 
 import pytest
 
+from soupsavvy.exceptions import NotSoupSelectorException, TagNotFoundException
 from soupsavvy.tags.combinators import DescendantCombinator, SelectorList
 from soupsavvy.tags.components import (
     AnyTagSelector,
@@ -9,7 +10,6 @@ from soupsavvy.tags.components import (
     PatternSelector,
     TagSelector,
 )
-from soupsavvy.tags.exceptions import NotSoupSelectorException, TagNotFoundException
 from tests.soupsavvy.tags.conftest import find_body_element, strip, to_bs
