+++ tests/soupsavvy/tags/components/tag_selector_test.py
@@ -6,8 +6,8 @@ import pytest
 from bs4 import Tag
 
 import soupsavvy.tags.namespace as ns
+from soupsavvy.exceptions import TagNotFoundException
 from soupsavvy.tags.components import AnyTagSelector, AttributeSelector, TagSelector
-from soupsavvy.tags.exceptions import TagNotFoundException
 from tests.soupsavvy.tags.conftest import (
     MockLinkSelector,
     find_body_element,
