+++ tests/soupsavvy/tags/css/nth/only_of_selector_test.py
@@ -2,8 +2,8 @@
 
 import pytest
 
+from soupsavvy.exceptions import NotSoupSelectorException, TagNotFoundException
 from soupsavvy.tags.css.nth.nth_soup_selector import OnlyOfSelector
-from soupsavvy.tags.exceptions import NotSoupSelectorException, TagNotFoundException
 from tests.soupsavvy.tags.conftest import (
     MockClassMenuSelector,
     MockDivSelector,
