+++ tests/soupsavvy/tags/css/tag_selectors/css_wrapper_test.py
@@ -5,8 +5,8 @@ for css selector-based search.
 
 import pytest
 
+from soupsavvy.exceptions import TagNotFoundException
 from soupsavvy.tags.css.tag_selectors import CSS
-from soupsavvy.tags.exceptions import TagNotFoundException
 from tests.soupsavvy.tags.conftest import find_body_element, strip, to_bs
