+++ tests/soupsavvy/tags/css/tag_selectors/only_of_type_test.py
@@ -2,8 +2,8 @@
 
 import pytest
 
+from soupsavvy.exceptions import TagNotFoundException
 from soupsavvy.tags.css.tag_selectors import OnlyOfType
-from soupsavvy.tags.exceptions import TagNotFoundException
 from tests.soupsavvy.tags.conftest import find_body_element, strip, to_bs
